Move network code to cmsis-update
