+++ rust/cmsis-update/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "cmsis-update"
+version = "0.1.0"
+authors = ["Jimmy Brisson <theotherjimmy@gmail.com>"]
+
+[dependencies]
+error-chain = "0.10.0"
+futures-await = "0.1.0"
+hyper = "0.11.2"
+hyper-tls = "0.1.2"
+minidom = "0.5.0"
+app_dirs = {git = "https://github.com/theotherjimmy/app-dirs-rs.git", branch = "fix-nightly-1-29"}
+clap = "2.19.0"
+slog-term = "^2"
+slog-async = "^2"
+tokio-core = "0.1.6"
+
+utils = { path = "../utils" }
+pack-index = { path = "../pack-index" }
+pdsc = { path = "../pdsc" }
